Refactor toot_handler.rs into smaller modules for better maintainability\n\nSplit the large toot_handler.rs file (1379 lines) into smaller, focused modules:\n- handler.rs: Main TootStreamHandler implementation\n- processor.rs: Core toot processing logic\n- coordinator.rs: Media recreation coordination\n- race.rs: Race condition handling\n- stats.rs: Processing statistics structures\n\nThis improves code organization, readability, and maintainability while preserving all existing functionality. All 340 tests continue to pass.
