--- src/toot_handler.rs
@@ -1,1379 +1,0 @@
-use crate::config::RuntimeConfig;
-use crate::error::{AlternatorError, MastodonError};
-use crate::language::LanguageDetector;
-use crate::mastodon::{MastodonClient, MastodonStream, MediaRecreation, TootEvent};
-use crate::media::MediaProcessor;
-use crate::openrouter::OpenRouterClient;
-use std::collections::HashSet;
-use tracing::{debug, error, info, warn};
-
-/// Handler for processing incoming toot events from WebSocket stream
-pub struct TootStreamHandler {
-    mastodon_client: MastodonClient,
-    openrouter_client: OpenRouterClient,
-    media_processor: MediaProcessor,
-    language_detector: LanguageDetector,
-    processed_toots: HashSet<String>,
-    config: RuntimeConfig,
-}
-
-impl TootStreamHandler {
-    /// Create a new toot stream handler
-    pub fn new(
-        mastodon_client: MastodonClient,
-        openrouter_client: OpenRouterClient,
-        media_processor: MediaProcessor,
-        language_detector: LanguageDetector,
-        config: RuntimeConfig,
-    ) -> Self {
-        Self {
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            processed_toots: HashSet::new(),
-            config,
-        }
-    }
-
-    /// Start processing toot stream - main entry point
-    pub async fn start_processing(&mut self) -> Result<(), AlternatorError> {
-        info!("Starting toot stream processing");
-
-        // Connect to Mastodon WebSocket stream
-        self.mastodon_client
-            .connect()
-            .await
-            .map_err(AlternatorError::Mastodon)?;
-
-        info!("✓ Connected to Mastodon stream - listening for toots");
-
-        // Main processing loop
-        loop {
-            match self.listen_and_process().await {
-                Ok(()) => {
-                    // Continue processing
-                    continue;
-                }
-                Err(e) => {
-                    error!("Error in toot processing loop: {}", e);
-
-                    // Handle specific error types
-                    match &e {
-                        AlternatorError::Mastodon(MastodonError::Disconnected(_))
-                        | AlternatorError::Mastodon(MastodonError::ConnectionFailed(_)) => {
-                            warn!("Connection lost, will attempt to reconnect");
-                            // The MastodonClient will handle reconnection automatically
-                            continue;
-                        }
-                        _ => {
-                            // For other errors, propagate up to main application
-                            return Err(e);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Listen for a single toot event and process it
-    async fn listen_and_process(&mut self) -> Result<(), AlternatorError> {
-        // Listen for toot events
-        match self.mastodon_client.listen().await {
-            Ok(Some(toot)) => {
-                // Verify this is from the authenticated user (already done in MastodonClient)
-                // Check for duplicate processing
-                if self.is_already_processed(&toot.id) {
-                    debug!("Skipping already processed toot: {}", toot.id);
-                    return Ok(());
-                }
-
-                info!(
-                    "Processing toot: {} (media: {})",
-                    toot.id,
-                    toot.media_attachments.len()
-                );
-
-                // Process the toot
-                match self.process_toot(&toot).await {
-                    Ok(()) => {
-                        self.mark_as_processed(toot.id.clone());
-                        info!("✓ Successfully processed toot: {}", toot.id);
-                    }
-                    Err(e) => {
-                        // Log error but continue processing other toots
-                        error!("Failed to process toot {}: {}", toot.id, e);
-
-                        // Still mark as processed to avoid retry loops for non-recoverable errors
-                        self.mark_as_processed(toot.id.clone());
-
-                        // Return error for recoverable issues that should be handled at higher level
-                        match &e {
-                            AlternatorError::Mastodon(MastodonError::RateLimitExceeded {
-                                ..
-                            })
-                            | AlternatorError::OpenRouter(
-                                crate::error::OpenRouterError::RateLimitExceeded { .. },
-                            ) => {
-                                return Err(e);
-                            }
-                            _ => {
-                                // For other errors, log and continue
-                                warn!(
-                                    "Non-recoverable error processing toot {}, continuing: {}",
-                                    toot.id, e
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(None) => {
-                // No toot received, continue listening
-                debug!("No toot received, continuing to listen");
-            }
-            Err(e) => {
-                error!("Error listening for toots: {}", e);
-                return Err(AlternatorError::Mastodon(e));
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Process a single toot - check for media, generate descriptions, and update
-    async fn process_toot(&self, toot: &TootEvent) -> Result<(), AlternatorError> {
-        // Check if toot has media attachments
-        if toot.media_attachments.is_empty() {
-            debug!("Toot {} has no media attachments, skipping", toot.id);
-            return Ok(());
-        }
-
-        // Filter media that needs processing (with audio awareness)
-        let processable_media = self.media_processor.filter_processable_media_with_audio(
-            &toot.media_attachments,
-            self.config.is_audio_enabled(),
-        );
-
-        if processable_media.is_empty() {
-            debug!(
-                "Toot {} has no processable media (all have descriptions or unsupported types)",
-                toot.id
-            );
-            return Ok(());
-        }
-
-        info!(
-            "Found {} processable media attachments in toot {}",
-            processable_media.len(),
-            toot.id
-        );
-
-        // Detect language for prompt selection
-        let detected_language = self.detect_toot_language(toot)?;
-        let prompt_template = self
-            .language_detector
-            .get_prompt_template(&detected_language)
-            .map_err(AlternatorError::Language)?;
-
-        debug!(
-            "Using language '{}' with prompt template",
-            detected_language
-        );
-
-        // Process each media attachment and collect successful descriptions with media data
-        let mut media_recreations: Vec<MediaRecreation> = Vec::new();
-        let mut original_media_ids = Vec::new();
-
-        // First pass: Prepare all media for processing (downloads and preprocessing)
-        let mut prepared_media = Vec::new();
-
-        for media in processable_media {
-            info!(
-                "Preparing media attachment: {} ({})",
-                media.id, media.media_type
-            );
-
-            // Check for race conditions before processing
-            if let Err(e) = self.check_race_condition(&toot.id, &media.id).await {
-                match e {
-                    AlternatorError::Mastodon(MastodonError::RaceConditionDetected) => {
-                        info!("Race condition detected for media {}, skipping", media.id);
-                        continue;
-                    }
-                    _ => {
-                        warn!(
-                            "Could not check race condition for media {}: {}",
-                            media.id, e
-                        );
-                        // Continue processing but log the warning
-                    }
-                }
-            }
-
-            // Handle audio files separately from images
-            if media.media_type.to_lowercase().starts_with("audio") {
-                // Check if audio processing is enabled
-                if !self.config.is_audio_enabled() {
-                    debug!(
-                        "Audio processing disabled, skipping audio file: {} ({})",
-                        media.id, media.media_type
-                    );
-                    continue;
-                }
-
-                info!("Processing audio file: {} ({})", media.id, media.media_type);
-
-                // Download original audio data for recreation
-                let original_audio_data = match self
-                    .media_processor
-                    .download_media_for_recreation(media)
-                    .await
-                {
-                    Ok(data) => data,
-                    Err(e) => {
-                        error!(
-                            "Failed to download audio {} for recreation: {}",
-                            media.id, e
-                        );
-                        continue;
-                    }
-                };
-
-                // Transcribe the audio to get description
-                match crate::media::process_audio_for_transcript(
-                    media,
-                    self.config.config.whisper(),
-                    self.config.config.media(),
-                    Some(&self.config.config.openrouter),
-                )
-                .await
-                {
-                    Ok(transcript) => {
-                        info!(
-                            "Generated transcript for audio {}: {}",
-                            media.id, transcript
-                        );
-                        // Determine appropriate file extension for audio
-                        let extension = match media.media_type.as_str() {
-                            "audio/mpeg" | "audio/mp3" => "mp3",
-                            "audio/wav" | "audio/wave" | "audio/x-wav" => "wav",
-                            "audio/m4a" => "m4a",
-                            "audio/mp4" => "mp4",
-                            "audio/aac" => "aac",
-                            "audio/ogg" => "ogg",
-                            "audio/flac" | "audio/x-flac" => "flac",
-                            _ => "audio", // fallback
-                        };
-                        let filename = format!("audio_{}.{}", media.id, extension);
-
-                        // Add to media recreations for batch processing
-                        media_recreations.push(MediaRecreation {
-                            data: original_audio_data,
-                            description: transcript,
-                            media_type: media.media_type.clone(),
-                            filename,
-                        });
-                        // Track original media ID for cleanup
-                        original_media_ids.push(media.id.clone());
-                    }
-                    Err(crate::error::MediaError::UnsupportedType { .. }) => {
-                        warn!(
-                            "Audio type {} not supported for transcription, skipping",
-                            media.media_type
-                        );
-                        continue;
-                    }
-                    Err(e) => {
-                        error!("Failed to transcribe audio {}: {}", media.id, e);
-                        continue;
-                    }
-                }
-
-                // Skip to next media (don't process as image)
-                continue;
-            }
-
-            // Handle video files separately from images
-            if media.media_type.to_lowercase().starts_with("video") {
-                debug!(
-                    "Detected video media: '{}' (type: '{}')",
-                    media.id, media.media_type
-                );
-
-                // Check if audio processing is enabled (required for video transcription)
-                if !self.config.is_audio_enabled() {
-                    debug!(
-                        "Audio processing disabled, skipping video file: {} ({})",
-                        media.id, media.media_type
-                    );
-                    continue;
-                }
-
-                info!("Processing video file: {} ({})", media.id, media.media_type);
-
-                // Download original video data for recreation
-                debug!(
-                    "About to download video data for media: {} with type: '{}'",
-                    media.id, media.media_type
-                );
-                let original_video_data = match self
-                    .media_processor
-                    .download_media_for_recreation(media)
-                    .await
-                {
-                    Ok(data) => data,
-                    Err(e) => {
-                        error!(
-                            "Failed to download video {} for recreation: {}",
-                            media.id, e
-                        );
-                        continue;
-                    }
-                };
-
-                // Transcribe the video audio to get description
-                match crate::media::process_video_for_transcript(
-                    media,
-                    self.config.config.whisper(),
-                    self.config.config.media(),
-                    Some(&self.config.config.openrouter),
-                )
-                .await
-                {
-                    Ok(transcript) => {
-                        info!(
-                            "Generated transcript for video {}: {}",
-                            media.id, transcript
-                        );
-                        // Determine appropriate file extension for video
-                        let extension = match media.media_type.as_str() {
-                            "video/mp4" => "mp4",
-                            "video/mpeg" => "mpeg",
-                            "video/quicktime" => "mov",
-                            "video/x-msvideo" => "avi",
-                            "video/webm" => "webm",
-                            "video/x-ms-wmv" => "wmv",
-                            "video/x-flv" => "flv",
-                            "video/3gpp" => "3gp",
-                            "video/x-matroska" => "mkv",
-                            _ => "video", // fallback
-                        };
-                        let filename = format!("video_{}.{}", media.id, extension);
-
-                        // Add to media recreations for batch processing
-                        media_recreations.push(MediaRecreation {
-                            data: original_video_data,
-                            description: transcript,
-                            media_type: media.media_type.clone(),
-                            filename,
-                        });
-                        // Track original media ID for cleanup
-                        original_media_ids.push(media.id.clone());
-                    }
-                    Err(crate::error::MediaError::UnsupportedType { .. }) => {
-                        warn!(
-                            "Video type {} not supported for transcription, skipping",
-                            media.media_type
-                        );
-                        continue;
-                    }
-                    Err(e) => {
-                        error!("Failed to transcribe video {}: {}", media.id, e);
-                        continue;
-                    }
-                }
-
-                // Skip to next media (don't process as image)
-                continue;
-            }
-
-            // Process non-audio media (images) through the existing pipeline
-            // Download original image data for recreation
-            let original_image_data = match self
-                .media_processor
-                .download_media_for_recreation(media)
-                .await
-            {
-                Ok(data) => data,
-                Err(e) => {
-                    error!(
-                        "Failed to download media {} for recreation: {}",
-                        media.id, e
-                    );
-                    continue; // Skip this media but continue with others
-                }
-            };
-
-            // Process media for analysis (resized/optimized version)
-            let processed_media_data =
-                match self.media_processor.process_media_for_analysis(media).await {
-                    Ok(data) => data,
-                    Err(e) => {
-                        error!("Failed to process media {} for analysis: {}", media.id, e);
-                        continue; // Skip this media but continue with others
-                    }
-                };
-
-            prepared_media.push((media.clone(), original_image_data, processed_media_data));
-        }
-
-        if prepared_media.is_empty() && media_recreations.is_empty() {
-            debug!("No media could be prepared for processing");
-            return Ok(());
-        }
-
-        // Process images if any are prepared
-        if !prepared_media.is_empty() {
-            info!(
-                "Prepared {} media attachments, starting parallel description generation",
-                prepared_media.len()
-            );
-
-            // Second pass: Generate descriptions in parallel using OpenRouter
-            let description_tasks: Vec<_> = prepared_media
-                .iter()
-                .map(|(media, _original_data, processed_data)| {
-                    let openrouter_client = &self.openrouter_client;
-                    let media_id = media.id.clone();
-                    async move {
-                        let result = openrouter_client
-                            .describe_image(processed_data, prompt_template)
-                            .await;
-                        (media_id, result)
-                    }
-                })
-                .collect();
-
-            let description_results = futures_util::future::join_all(description_tasks).await;
-
-            // Process results and build media recreations
-            for ((media, original_data, _processed_data), (result_media_id, description_result)) in
-                prepared_media
-                    .into_iter()
-                    .zip(description_results.into_iter())
-            {
-                debug_assert_eq!(
-                    media.id, result_media_id,
-                    "Media ID mismatch in parallel processing"
-                );
-
-                match description_result {
-                    Ok(description) => {
-                        info!(
-                            "Generated description for media {}: {}",
-                            media.id, description
-                        );
-
-                        // Determine appropriate file extension for images
-                        let extension = match media.media_type.as_str() {
-                            "image/jpeg" => "jpg",
-                            "image/png" => "png",
-                            "image/gif" => "gif",
-                            "image/webp" => "webp",
-                            "image/bmp" => "bmp",
-                            "image/tiff" => "tiff",
-                            _ => "jpg", // fallback to jpg for unknown image types
-                        };
-                        let filename = format!("image_{}.{}", media.id, extension);
-
-                        // Add to media recreations for batch processing
-                        media_recreations.push(MediaRecreation {
-                            data: original_data,
-                            description,
-                            media_type: media.media_type.clone(),
-                            filename,
-                        });
-                        // Track original media ID for cleanup
-                        original_media_ids.push(media.id);
-                    }
-                    Err(crate::error::OpenRouterError::TokenLimitExceeded { .. }) => {
-                        warn!("Token limit exceeded for media {}, skipping", media.id);
-                        continue; // Skip this media but continue with others
-                    }
-                    Err(e) => {
-                        error!(
-                            "Failed to generate description for media {}: {}",
-                            media.id, e
-                        );
-                        return Err(AlternatorError::OpenRouter(e));
-                    }
-                }
-            }
-        }
-
-        // If we have media to recreate, do a batch recreation
-        if !media_recreations.is_empty() {
-            info!(
-                "Recreating {} media attachments with descriptions for toot {}",
-                media_recreations.len(),
-                toot.id
-            );
-
-            // Final race condition check and batch recreation
-            match self
-                .recreate_media_with_race_check(
-                    &toot.id,
-                    media_recreations.clone(),
-                    original_media_ids.clone(),
-                )
-                .await
-            {
-                Ok(()) => {
-                    info!(
-                        "✓ Successfully recreated {} media attachments for toot: {}",
-                        media_recreations.len(),
-                        toot.id
-                    );
-                }
-                Err(AlternatorError::Mastodon(MastodonError::RaceConditionDetected)) => {
-                    info!(
-                        "Race condition detected during media recreation for toot {}, operation aborted",
-                        toot.id
-                    );
-                }
-                Err(e) => {
-                    error!(
-                        "Failed to recreate media attachments for toot {}: {}",
-                        toot.id, e
-                    );
-                    return Err(e);
-                }
-            }
-        } else {
-            info!("No media attachments to recreate for toot {}", toot.id);
-        }
-
-        Ok(())
-    }
-
-    /// Detect the language of a toot with fallback handling
-    #[allow(clippy::result_large_err)] // AlternatorError is large but needed for comprehensive error handling
-    fn detect_toot_language(&self, toot: &TootEvent) -> Result<String, AlternatorError> {
-        // First, check if the toot has a language attribute
-        if let Some(ref lang) = toot.language {
-            if !lang.trim().is_empty() {
-                debug!("Using toot language attribute: {}", lang);
-                return Ok(lang.clone());
-            }
-        }
-
-        // Fallback to content-based language detection
-        debug!("No toot language attribute found, detecting from content");
-        match self.language_detector.detect_language(&toot.content) {
-            Ok(lang) => {
-                debug!("Detected language from content: {}", lang);
-                Ok(lang)
-            }
-            Err(e) => {
-                warn!("Language detection failed: {}, defaulting to English", e);
-                Ok("en".to_string())
-            }
-        }
-    }
-
-    /// Check for race conditions by retrieving current toot state
-    async fn check_race_condition(
-        &self,
-        toot_id: &str,
-        media_id: &str,
-    ) -> Result<(), AlternatorError> {
-        debug!(
-            "Checking for race conditions on toot {} media {}",
-            toot_id, media_id
-        );
-
-        match self.mastodon_client.get_toot(toot_id).await {
-            Ok(current_toot) => {
-                // Find the current state of this media attachment
-                if let Some(current_media) = current_toot
-                    .media_attachments
-                    .iter()
-                    .find(|m| m.id == *media_id)
-                {
-                    if current_media.description.is_some()
-                        && !current_media
-                            .description
-                            .as_ref()
-                            .unwrap()
-                            .trim()
-                            .is_empty()
-                    {
-                        debug!(
-                            "Media {} already has description, race condition detected",
-                            media_id
-                        );
-                        return Err(AlternatorError::Mastodon(
-                            MastodonError::RaceConditionDetected,
-                        ));
-                    }
-                } else {
-                    // Media attachment not found in current toot state
-                    debug!(
-                        "Media {} no longer exists in toot {}, race condition detected",
-                        media_id, toot_id
-                    );
-                    return Err(AlternatorError::Mastodon(
-                        MastodonError::RaceConditionDetected,
-                    ));
-                }
-                Ok(())
-            }
-            Err(e) => {
-                warn!(
-                    "Could not retrieve current toot state for race condition check: {}",
-                    e
-                );
-                Err(AlternatorError::Mastodon(e))
-            }
-        }
-    }
-
-    /// Update media description with a final race condition check
-    #[allow(dead_code)] // Kept for backward compatibility, replaced by batch update
-    async fn update_media_with_race_check(
-        &self,
-        toot_id: &str,
-        media_id: &str,
-        description: &str,
-    ) -> Result<(), AlternatorError> {
-        // Final race condition check before update
-        self.check_race_condition(toot_id, media_id).await?;
-
-        // Update media description
-        match self
-            .mastodon_client
-            .update_media(toot_id, media_id, description)
-            .await
-        {
-            Ok(()) => Ok(()),
-            Err(MastodonError::MediaNotFound { .. }) => {
-                // Treat MediaNotFound as a race condition - the media was removed/changed
-                debug!(
-                    "Media {} not found during update, treating as race condition",
-                    media_id
-                );
-                Err(AlternatorError::Mastodon(
-                    MastodonError::RaceConditionDetected,
-                ))
-            }
-            Err(e) => Err(AlternatorError::Mastodon(e)),
-        }
-    }
-
-    /// Recreate media attachments with descriptions and race condition checks
-    async fn recreate_media_with_race_check(
-        &self,
-        toot_id: &str,
-        media_recreations: Vec<MediaRecreation>, // Vec of media recreations with descriptions
-        original_media_ids: Vec<String>,         // Original media IDs to clean up after success
-    ) -> Result<(), AlternatorError> {
-        if media_recreations.is_empty() {
-            return Ok(());
-        }
-
-        // Get current toot state to verify no race conditions
-        let current_toot = self
-            .mastodon_client
-            .get_toot(toot_id)
-            .await
-            .map_err(AlternatorError::Mastodon)?;
-
-        // Check if any of the original media attachments now have descriptions
-        let processable_media = self
-            .media_processor
-            .filter_processable_media(&current_toot.media_attachments);
-
-        if processable_media.len() != media_recreations.len() {
-            debug!(
-                "Media state changed: expected {} processable media, found {}. Race condition detected.",
-                media_recreations.len(),
-                processable_media.len()
-            );
-            return Err(AlternatorError::Mastodon(
-                MastodonError::RaceConditionDetected,
-            ));
-        }
-
-        // Recreate all media attachments with descriptions (includes cleanup)
-        match self
-            .mastodon_client
-            .recreate_media_with_descriptions(toot_id, media_recreations, original_media_ids)
-            .await
-        {
-            Ok(()) => Ok(()),
-            Err(e) => Err(AlternatorError::Mastodon(e)),
-        }
-    }
-
-    /// Check if a toot has already been processed
-    fn is_already_processed(&self, toot_id: &str) -> bool {
-        self.processed_toots.contains(toot_id)
-    }
-
-    /// Mark a toot as processed to prevent duplicate processing
-    fn mark_as_processed(&mut self, toot_id: String) {
-        self.processed_toots.insert(toot_id);
-
-        // Prevent memory growth by limiting the size of processed toots set
-        if self.processed_toots.len() > 10000 {
-            // Remove oldest entries (this is a simple approach, could be improved with LRU)
-            let excess = self.processed_toots.len() - 5000;
-            let to_remove: Vec<String> =
-                self.processed_toots.iter().take(excess).cloned().collect();
-            for id in to_remove {
-                self.processed_toots.remove(&id);
-            }
-            debug!(
-                "Cleaned up processed toots cache, now contains {} entries",
-                self.processed_toots.len()
-            );
-        }
-    }
-
-    /// Get statistics about processed toots
-    #[allow(dead_code)] // Public API method, may be used in future
-    pub fn get_processing_stats(&self) -> ProcessingStats {
-        ProcessingStats {
-            processed_toots_count: self.processed_toots.len(),
-        }
-    }
-
-    /// Clear the processed toots cache (useful for testing)
-    #[cfg(test)]
-    pub fn clear_processed_cache(&mut self) {
-        self.processed_toots.clear();
-    }
-}
-
-/// Statistics about toot processing
-#[allow(dead_code)] // Stats struct for API completeness
-#[derive(Debug, Clone)]
-pub struct ProcessingStats {
-    pub processed_toots_count: usize,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::{
-        BalanceConfig, Config, LoggingConfig, MastodonConfig, MediaConfig, OpenRouterConfig,
-        RuntimeConfig, WhisperConfig,
-    };
-    use crate::mastodon::{Account, MediaAttachment, MediaDimensions, MediaMeta};
-    use chrono::Utc;
-
-    fn create_test_mastodon_config() -> MastodonConfig {
-        MastodonConfig {
-            instance_url: "https://mastodon.social".to_string(),
-            access_token: "test_token".to_string(),
-            user_stream: Some(true),
-        }
-    }
-
-    fn create_test_openrouter_config() -> OpenRouterConfig {
-        OpenRouterConfig {
-            api_key: "test_key".to_string(),
-            model: "mistralai/mistral-small-3.2-24b-instruct:free".to_string(),
-            vision_model: "mistralai/mistral-small-3.2-24b-instruct:free".to_string(),
-            text_model: "mistralai/mistral-small-3.2-24b-instruct:free".to_string(),
-            base_url: Some("https://test.openrouter.ai/api/v1".to_string()),
-            max_tokens: Some(150),
-        }
-    }
-
-    fn create_test_runtime_config() -> RuntimeConfig {
-        let config = Config {
-            mastodon: create_test_mastodon_config(),
-            openrouter: create_test_openrouter_config(),
-            media: Some(MediaConfig::default()),
-            balance: Some(BalanceConfig::default()),
-            logging: Some(LoggingConfig::default()),
-            whisper: Some(WhisperConfig::default()),
-        };
-        RuntimeConfig::new(config)
-    }
-
-    #[allow(dead_code)] // Test helper function
-    fn create_test_toot(id: &str, media_attachments: Vec<MediaAttachment>) -> TootEvent {
-        TootEvent {
-            id: id.to_string(),
-            uri: format!("https://mastodon.social/users/testuser/statuses/{id}"),
-            account: Account {
-                id: "user123".to_string(),
-                username: "testuser".to_string(),
-                acct: "testuser@mastodon.social".to_string(),
-                display_name: "Test User".to_string(),
-                url: "https://mastodon.social/@testuser".to_string(),
-            },
-            content: "Test toot with image".to_string(),
-            language: Some("en".to_string()),
-            media_attachments,
-            created_at: Utc::now(),
-            url: Some("https://mastodon.social/@testuser/123456789".to_string()),
-            visibility: "public".to_string(),
-            sensitive: false,
-            spoiler_text: String::new(),
-            in_reply_to_id: None,
-            in_reply_to_account_id: None,
-            mentions: Vec::new(),
-            tags: Vec::new(),
-            emojis: Vec::new(),
-            poll: None,
-        }
-    }
-
-    #[allow(dead_code)] // Test helper function
-    fn create_test_media(
-        id: &str,
-        media_type: &str,
-        description: Option<String>,
-    ) -> MediaAttachment {
-        MediaAttachment {
-            id: id.to_string(),
-            media_type: media_type.to_string(),
-            url: format!("https://example.com/media/{id}"),
-            preview_url: None,
-            description,
-            meta: None,
-        }
-    }
-
-    #[test]
-    fn test_toot_stream_handler_creation() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        assert_eq!(handler.processed_toots.len(), 0);
-    }
-
-    #[test]
-    fn test_duplicate_processing_prevention() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Initially not processed
-        assert!(!handler.is_already_processed("toot123"));
-
-        // Mark as processed
-        handler.mark_as_processed("toot123".to_string());
-        assert!(handler.is_already_processed("toot123"));
-
-        // Different toot should not be marked
-        assert!(!handler.is_already_processed("toot456"));
-    }
-
-    #[test]
-    fn test_processed_cache_cleanup() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Add many processed toots to trigger cleanup
-        for i in 0..10001 {
-            handler.mark_as_processed(format!("toot{i}"));
-        }
-
-        // Should have triggered cleanup
-        assert!(handler.processed_toots.len() <= 5000);
-    }
-
-    #[test]
-    fn test_language_detection_fallback() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Test with toot that has language attribute
-        let toot_with_lang = create_test_toot("123", vec![]);
-        let mut toot_with_lang = toot_with_lang;
-        toot_with_lang.language = Some("de".to_string());
-
-        let result = handler.detect_toot_language(&toot_with_lang).unwrap();
-        assert_eq!(result, "de");
-
-        // Test with toot that has empty language attribute (should fallback to content detection)
-        let mut toot_empty_lang = create_test_toot("456", vec![]);
-        toot_empty_lang.language = Some("".to_string());
-        toot_empty_lang.content = "Hello world, this is a test".to_string();
-
-        let result = handler.detect_toot_language(&toot_empty_lang).unwrap();
-        assert_eq!(result, "en");
-
-        // Test with toot without language attribute (should fallback to content detection)
-        let mut toot_no_lang = create_test_toot("789", vec![]);
-        toot_no_lang.language = None;
-        toot_no_lang.content = "Das ist ein Test mit deutschen Wörtern".to_string();
-
-        let result = handler.detect_toot_language(&toot_no_lang).unwrap();
-        assert_eq!(result, "de");
-
-        // Test with toot without language and empty content (should fallback to English)
-        let mut toot_empty = create_test_toot("000", vec![]);
-        toot_empty.language = None;
-        toot_empty.content = "".to_string();
-
-        let result = handler.detect_toot_language(&toot_empty).unwrap();
-        assert_eq!(result, "en");
-    }
-
-    #[test]
-    fn test_processing_stats() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let stats = handler.get_processing_stats();
-        assert_eq!(stats.processed_toots_count, 0);
-
-        handler.mark_as_processed("toot1".to_string());
-        handler.mark_as_processed("toot2".to_string());
-
-        let stats = handler.get_processing_stats();
-        assert_eq!(stats.processed_toots_count, 2);
-    }
-
-    #[test]
-    fn test_clear_processed_cache() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        handler.mark_as_processed("toot1".to_string());
-        handler.mark_as_processed("toot2".to_string());
-        assert_eq!(handler.processed_toots.len(), 2);
-
-        handler.clear_processed_cache();
-        assert_eq!(handler.processed_toots.len(), 0);
-    }
-
-    #[test]
-    fn test_processing_stats_debug() {
-        let stats = ProcessingStats {
-            processed_toots_count: 42,
-        };
-
-        let debug_str = format!("{stats:?}");
-        assert!(debug_str.contains("processed_toots_count: 42"));
-    }
-
-    #[tokio::test]
-    async fn test_process_toot_no_media() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let toot = create_test_toot("123", vec![]);
-        let result = handler.process_toot(&toot).await;
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_detect_toot_language_with_attribute() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let mut toot = create_test_toot("123", vec![]);
-        toot.language = Some("fr".to_string());
-
-        let result = handler.detect_toot_language(&toot).unwrap();
-        assert_eq!(result, "fr");
-    }
-
-    #[test]
-    fn test_detect_toot_language_with_whitespace_attribute() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let mut toot = create_test_toot("123", vec![]);
-        toot.language = Some("  ".to_string());
-        toot.content = "Ciao mondo, questo è un test in italiano".to_string();
-
-        let result = handler.detect_toot_language(&toot).unwrap();
-        assert_eq!(result, "it");
-    }
-
-    #[test]
-    fn test_detect_toot_language_fallback_to_english() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let mut toot = create_test_toot("123", vec![]);
-        toot.language = None;
-        toot.content = "".to_string();
-
-        // Should fallback to English when detection fails
-        let result = handler.detect_toot_language(&toot).unwrap();
-        assert_eq!(result, "en");
-    }
-
-    #[test]
-    fn test_media_attachment_creation() {
-        let media = create_test_media("media123", "image", None);
-        assert_eq!(media.id, "media123");
-        assert_eq!(media.media_type, "image");
-        assert!(media.description.is_none());
-        assert_eq!(media.url, "https://example.com/media/media123");
-
-        let media_with_desc = create_test_media("media456", "video", Some("A video".to_string()));
-        assert_eq!(media_with_desc.description, Some("A video".to_string()));
-    }
-
-    #[test]
-    fn test_toot_event_creation() {
-        let media = vec![create_test_media("media1", "image", None)];
-        let toot = create_test_toot("toot123", media.clone());
-
-        assert_eq!(toot.id, "toot123");
-        assert_eq!(toot.account.username, "testuser");
-        assert_eq!(toot.content, "Test toot with image");
-        assert_eq!(toot.language, Some("en".to_string()));
-        assert_eq!(toot.media_attachments.len(), 1);
-        assert_eq!(toot.media_attachments[0].id, "media1");
-        assert_eq!(toot.visibility, "public");
-    }
-
-    #[test]
-    fn test_account_fields() {
-        let toot = create_test_toot("123", vec![]);
-        let account = &toot.account;
-
-        assert_eq!(account.id, "user123");
-        assert_eq!(account.username, "testuser");
-        assert_eq!(account.acct, "testuser@mastodon.social");
-        assert_eq!(account.display_name, "Test User");
-        assert_eq!(account.url, "https://mastodon.social/@testuser");
-    }
-
-    #[test]
-    fn test_handler_stats_increment() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        assert_eq!(handler.get_processing_stats().processed_toots_count, 0);
-
-        handler.mark_as_processed("toot1".to_string());
-        assert_eq!(handler.get_processing_stats().processed_toots_count, 1);
-
-        handler.mark_as_processed("toot2".to_string());
-        handler.mark_as_processed("toot3".to_string());
-        assert_eq!(handler.get_processing_stats().processed_toots_count, 3);
-    }
-
-    #[test]
-    fn test_duplicate_toot_handling() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Mark a toot as processed multiple times
-        handler.mark_as_processed("toot123".to_string());
-        handler.mark_as_processed("toot123".to_string());
-        handler.mark_as_processed("toot123".to_string());
-
-        // Should only count once
-        assert_eq!(handler.get_processing_stats().processed_toots_count, 1);
-        assert!(handler.is_already_processed("toot123"));
-    }
-
-    #[test]
-    fn test_cache_size_management() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let mut handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Test that cache management works properly
-        for i in 0..100 {
-            handler.mark_as_processed(format!("toot{i}"));
-        }
-
-        assert_eq!(handler.get_processing_stats().processed_toots_count, 100);
-
-        // Cache should still contain all items for reasonable sizes
-        for i in 0..100 {
-            assert!(handler.is_already_processed(&format!("toot{i}")));
-        }
-    }
-
-    #[test]
-    fn test_language_detection_priority() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        // Test that toot language attribute takes priority over content detection
-        let mut toot = create_test_toot("123", vec![]);
-        toot.language = Some("es".to_string());
-        toot.content = "Hello world this is English text".to_string();
-
-        let result = handler.detect_toot_language(&toot).unwrap();
-        assert_eq!(result, "es"); // Should use toot attribute, not content detection
-    }
-
-    #[test]
-    fn test_media_meta_creation() {
-        let media = MediaAttachment {
-            id: "media_test".to_string(),
-            media_type: "image".to_string(),
-            url: "https://example.com/test.jpg".to_string(),
-            preview_url: Some("https://example.com/preview.jpg".to_string()),
-            description: None,
-            meta: Some(MediaMeta {
-                original: Some(MediaDimensions {
-                    width: Some(1920),
-                    height: Some(1080),
-                    size: Some("1920x1080".to_string()),
-                    aspect: Some(1.777),
-                }),
-                small: Some(MediaDimensions {
-                    width: Some(400),
-                    height: Some(225),
-                    size: Some("400x225".to_string()),
-                    aspect: Some(1.777),
-                }),
-            }),
-        };
-
-        assert_eq!(media.id, "media_test");
-        assert!(media.preview_url.is_some());
-        assert!(media.meta.is_some());
-
-        let meta = media.meta.unwrap();
-        assert!(meta.original.is_some());
-        assert!(meta.small.is_some());
-
-        let original = meta.original.unwrap();
-        assert_eq!(original.width, Some(1920));
-        assert_eq!(original.height, Some(1080));
-        assert_eq!(original.aspect, Some(1.777));
-    }
-
-    #[test]
-    fn test_processing_stats_clone() {
-        let stats = ProcessingStats {
-            processed_toots_count: 15,
-        };
-
-        let cloned_stats = stats.clone();
-        assert_eq!(
-            stats.processed_toots_count,
-            cloned_stats.processed_toots_count
-        );
-    }
-
-    #[test]
-    fn test_toot_event_serialization() {
-        let media = vec![create_test_media(
-            "media1",
-            "image",
-            Some("Test image".to_string()),
-        )];
-        let toot = create_test_toot("toot123", media);
-
-        // Test that the toot can be serialized (important for debugging and logging)
-        let serialized = serde_json::to_string(&toot);
-        assert!(serialized.is_ok());
-
-        let json_str = serialized.unwrap();
-        assert!(json_str.contains("toot123"));
-        assert!(json_str.contains("media1"));
-        assert!(json_str.contains("Test image"));
-    }
-
-    #[test]
-    fn test_account_serialization() {
-        let toot = create_test_toot("123", vec![]);
-        let account = &toot.account;
-
-        let serialized = serde_json::to_string(account);
-        assert!(serialized.is_ok());
-
-        let json_str = serialized.unwrap();
-        assert!(json_str.contains("testuser"));
-        assert!(json_str.contains("Test User"));
-    }
-
-    #[test]
-    fn test_media_attachment_serialization() {
-        let media = create_test_media("media123", "video", Some("A test video".to_string()));
-
-        let serialized = serde_json::to_string(&media);
-        assert!(serialized.is_ok());
-
-        let json_str = serialized.unwrap();
-        assert!(json_str.contains("media123"));
-        assert!(json_str.contains("video"));
-        assert!(json_str.contains("A test video"));
-    }
-
-    #[test]
-    fn test_empty_content_language_detection() {
-        let mastodon_client = MastodonClient::new(create_test_mastodon_config());
-        let openrouter_client = OpenRouterClient::new(create_test_openrouter_config());
-        let media_processor = MediaProcessor::with_default_config();
-        let language_detector = LanguageDetector::new();
-        let _config = create_test_runtime_config();
-
-        let handler = TootStreamHandler::new(
-            mastodon_client,
-            openrouter_client,
-            media_processor,
-            language_detector,
-            _config,
-        );
-
-        let mut toot = create_test_toot("123", vec![]);
-        toot.language = None;
-        toot.content = "".to_string();
-
-        // Should fallback to English for empty content
-        let result = handler.detect_toot_language(&toot).unwrap();
-        assert_eq!(result, "en");
-    }
-}
