+++ src/toot_handler/handler.rs
@@ -1,0 +1,185 @@
+use crate::config::RuntimeConfig;
+use crate::error::{AlternatorError, MastodonError};
+use crate::language::LanguageDetector;
+use crate::mastodon::{MastodonClient, MastodonStream};
+use crate::media::MediaProcessor;
+use crate::openrouter::OpenRouterClient;
+use crate::toot_handler::processor;
+use crate::toot_handler::stats::ProcessingStats;
+use std::collections::HashSet;
+use tracing::{debug, error, info, warn};
+
+/// Handler for processing incoming toot events from WebSocket stream
+pub struct TootStreamHandler {
+    mastodon_client: MastodonClient,
+    openrouter_client: OpenRouterClient,
+    media_processor: MediaProcessor,
+    language_detector: LanguageDetector,
+    processed_toots: HashSet<String>,
+    config: RuntimeConfig,
+}
+
+impl TootStreamHandler {
+    /// Create a new toot stream handler
+    pub fn new(
+        mastodon_client: MastodonClient,
+        openrouter_client: OpenRouterClient,
+        media_processor: MediaProcessor,
+        language_detector: LanguageDetector,
+        config: RuntimeConfig,
+    ) -> Self {
+        Self {
+            mastodon_client,
+            openrouter_client,
+            media_processor,
+            language_detector,
+            processed_toots: HashSet::new(),
+            config,
+        }
+    }
+
+    /// Start processing toot stream - main entry point
+    pub async fn start_processing(&mut self) -> Result<(), AlternatorError> {
+        info!("Starting toot stream processing");
+
+        // Connect to Mastodon WebSocket stream
+        self.mastodon_client
+            .connect()
+            .await
+            .map_err(AlternatorError::Mastodon)?;
+
+        info!("✓ Connected to Mastodon stream - listening for toots");
+
+        // Main processing loop
+        loop {
+            match self.listen_and_process().await {
+                Ok(()) => {
+                    // Continue processing
+                    continue;
+                }
+                Err(e) => {
+                    error!("Error in toot processing loop: {}", e);
+
+                    // Handle specific error types
+                    match &e {
+                        AlternatorError::Mastodon(MastodonError::Disconnected(_))
+                        | AlternatorError::Mastodon(MastodonError::ConnectionFailed(_)) => {
+                            warn!("Connection lost, will attempt to reconnect");
+                            // The MastodonClient will handle reconnection automatically
+                            continue;
+                        }
+                        _ => {
+                            // For other errors, propagate up to main application
+                            return Err(e);
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    /// Listen for a single toot event and process it
+    async fn listen_and_process(&mut self) -> Result<(), AlternatorError> {
+        // Listen for toot events
+        match self.mastodon_client.listen().await {
+            Ok(Some(toot)) => {
+                // Verify this is from the authenticated user (already done in MastodonClient)
+                // Check for duplicate processing
+                if self.is_already_processed(&toot.id) {
+                    debug!("Skipping already processed toot: {}", toot.id);
+                    return Ok(());
+                }
+
+                info!(
+                    "Processing toot: {} (media: {})",
+                    toot.id,
+                    toot.media_attachments.len()
+                );
+
+                // Process the toot
+                match processor::process_toot(
+                    &toot,
+                    &self.mastodon_client,
+                    &self.openrouter_client,
+                    &self.media_processor,
+                    &self.language_detector,
+                    &self.config,
+                ).await {
+                    Ok(()) => {
+                        self.mark_as_processed(toot.id.clone());
+                        info!("✓ Successfully processed toot: {}", toot.id);
+                    }
+                    Err(e) => {
+                        // Log error but continue processing other toots
+                        error!("Failed to process toot {}: {}", toot.id, e);
+
+                        // Still mark as processed to avoid retry loops for non-recoverable errors
+                        self.mark_as_processed(toot.id.clone());
+
+                        // Return error for recoverable issues that should be handled at higher level
+                        match &e {
+                            AlternatorError::Mastodon(MastodonError::RateLimitExceeded {
+                                ..
+                            })
+                            | AlternatorError::OpenRouter(
+                                crate::error::OpenRouterError::RateLimitExceeded { .. },
+                            ) => {
+                                return Err(e);
+                            }
+                            _ => {
+                                // For other errors, log and continue
+                                warn!(
+                                    "Non-recoverable error processing toot {}, continuing: {}",
+                                    toot.id, e
+                                );
+                            }
+                        }
+                    }
+                }
+            }
+            Ok(None) => {
+                // No toot received, continue listening
+                debug!("No toot received, continuing to listen");
+            }
+            Err(e) => {
+                error!("Error listening for toots: {}", e);
+                return Err(AlternatorError::Mastodon(e));
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Check if a toot has already been processed
+    fn is_already_processed(&self, toot_id: &str) -> bool {
+        self.processed_toots.contains(toot_id)
+    }
+
+    /// Mark a toot as processed to prevent duplicate processing
+    fn mark_as_processed(&mut self, toot_id: String) {
+        self.processed_toots.insert(toot_id);
+
+        // Prevent memory growth by limiting the size of processed toots set
+        if self.processed_toots.len() > 10000 {
+            // Remove oldest entries (this is a simple approach, could be improved with LRU)
+            let excess = self.processed_toots.len() - 5000;
+            let to_remove: Vec<String> =
+                self.processed_toots.iter().take(excess).cloned().collect();
+            for id in to_remove {
+                self.processed_toots.remove(&id);
+            }
+            debug!(
+                "Cleaned up processed toots cache, now contains {} entries",
+                self.processed_toots.len()
+            );
+        }
+    }
+
+    /// Get statistics about processed toots
+    #[allow(dead_code)] // Public API method, may be used in future
+    pub fn get_processing_stats(&self) -> ProcessingStats {
+        ProcessingStats {
+            processed_toots_count: self.processed_toots.len(),
+        }
+    }
+}
